feat: move core types to root shared folder
