+++ shared/types/core.ts
@@ -1,0 +1,8 @@
+export interface ErrorDto {
+  message: string
+}
+
+export interface PrivateFormErrorDto extends ErrorDto {
+  isPageFound: true
+  formTitle: string
+}
