+++ src/types/api/core.ts
@@ -1,8 +1,1 @@
-export interface ErrorDto {
-  message: string
-}
-
-export interface PrivateFormErrorDto extends ErrorDto {
-  isPageFound: true
-  formTitle: string
-}
+export * from '../../../shared/types/core'
